Add padded combine function
`combineLs` was a naive function that shorted out when one of the
lists terminated. It is now a wrapper for `combineLsPadded`, which,
though ugly af, is a proper implementation of what I want with my
lists: combine them and pad them, so that the inner lists are always
the same length. This makes it trivial to encode back into CSV and
produce a proper table without all the cells losing their
position.

Functionality first, refactoring later.