Use command line arguments as input
Instead of using hard-coded filenames, I can provide argument names
via the command line. The number of arguments is not limited, and they
are treated as a list. I'm using maps and folds in `main` to reflect
that change.

Sadly, `stack build` is currently failing, and I'm not sure why.