Add CSV diffing function
I start with a hard-coded 'original' file, and diff all args with it.
The merging algorithm adds the args to the 'original' file one by one.
The keys that exist in the original but not in the appended file are
sent to the diff files. The keys that don't exist in the original
are dropped.